requirements.txt
pandas
scikit-learn
numpy
scipy
icrawler
tensorflow==2.1.0
opencv-python
face_recognition
Pillow
filterpy
flask
flask-cors
flask_restplus
rdflib
uuid
SPARQLWrapper
requests
pyyaml
pymongo